release(html,deps): Safari first public release v1.0.2, update web-eid-authtoken-validation-java to v1.2.0 and web-eid.js to v1.0.1
