+++ example/pom.xml
@@ -20,7 +20,7 @@
 		<java.version>1.8</java.version>
 		<maven-surefire-plugin.version>2.22.1</maven-surefire-plugin.version>
 		<caffeine.version>2.8.5</caffeine.version>
-		<webeid.version>1.1.0</webeid.version>
+		<webeid.version>1.2.0</webeid.version>
 		<jmockit.version>1.44</jmockit.version>
 	</properties>
