+++ example/src/main/resources/templates/welcome-with-file-upload-support.html
@@ -37,7 +37,7 @@
 
         <script type="module">
             "use strict";
-            import * as webeid from "/js/web-eid-1.0.0.js";
+            import * as webeid from "/js/web-eid-1.0.1.js";
             import { hideErrorMessage, showErrorMessage } from "/js/error-message.js";
 
             const signButton = document.querySelector("#webeid-sign-button");
